search_properties/requirements.txt
# Core
mysql-connector-python==8.0.32
requests==2.28.2

# testing
pytest==7.2.2
pytest-cov
faker==18.2.0

# formatting
isort
black
flake8

# debugging
ipython
pdbpp